requirements.txt
# Core Azure SDKs
azure-identity>=1.15.0
azure-storage-blob>=12.19.0
azure-data-tables>=12.5.0
azure-search-documents>=11.4.0
azure-functions>=1.18.0
azure-durable-functions>=1.2.9
msgraph-sdk>=1.0.0

# LLM & Embeddings
openai>=1.12.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1
tiktoken>=0.6.0

# Document Processing
PyMuPDF>=1.23.0  # fitz
python-docx>=1.1.0
python-pptx>=0.6.23
pandas>=2.2.0
openpyxl>=3.1.2

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
httpx>=0.26.0
aiohttp>=3.9.0

# Utilities
python-multipart>=0.0.9
pytz>=2024.1
tabulate>=0.9.0  # For markdown tables

# Monitoring & Logging
azure-monitor-opentelemetry>=1.2.0
opencensus-ext-azure>=1.1.13

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.1.0
flake8>=7.0.0
mypy>=1.8.0
